Add v3.1.1

+++ src/main/java/com/miltrainApp/service/impl/TrainingServiceImpl.java
@@ -9,6 +9,7 @@ import com.miltrainApp.exceptions.UserNotFoundException;
 import com.miltrainApp.repository.TrainingRepository;
 import com.miltrainApp.repository.UserRepository;
 import com.miltrainApp.service.TrainingService;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -22,9 +23,7 @@ public class TrainingServiceImpl implements TrainingService {
 
     private final UserRepository userRepository;
 
-
     private final TrainingRepository trainingRepository;
-
 
     public TrainingServiceImpl(UserRepository userRepository, TrainingRepository trainingRepository) {
         this.userRepository = userRepository;
@@ -38,8 +37,15 @@ public class TrainingServiceImpl implements TrainingService {
         return auth.getName();
     }
 
+    public Long getCurrentUserId() {
+        String login = getCurrentLogin();
+        return userRepository.findByLogin(login)
+                             .orElseThrow(() -> new UserNotFoundException("User not found"))
+                             .getId();
+    }
+
     @Override
-    @PreAuthorize("hasRole('USER')")
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     public Training createTraining(CreateTrainingRequestDTO createTrainingRequest) {
 
         String login = getCurrentLogin();
@@ -55,24 +61,57 @@ public class TrainingServiceImpl implements TrainingService {
     }
 
     @Override
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     public Training getTrainingById(Long trainingId) {
-        return trainingRepository.findById(trainingId)
-                                 .orElseThrow(() -> new TrainingNotFoundException("Training not found!"));
+
+        Training training = trainingRepository.findById(trainingId)
+                                              .orElseThrow(() -> new TrainingNotFoundException("Training not found!"));
+        Authentication auth = SecurityContextHolder.getContext()
+                                                   .getAuthentication();
+
+        boolean isUser = auth.getAuthorities()
+                             .stream()
+                             .anyMatch(a -> a.getAuthority()
+                                             .equals("ROLE_USER"));
+
+        if (isUser && !training.getUserId()
+                               .equals(getCurrentUserId())) {
+            throw new AccessDeniedException("You can only access your on trainings!");
+        }
+        return training;
     }
 
     @Override
-    public List<Training> getAllTrainingsByUser(Long userId) {
+    @PreAuthorize("hasRole('USER')")
+    public List<Training> getAllTrainingsByUser() {
+        Long userId = getCurrentUserId();
         return trainingRepository.findAllByUserId(userId);
     }
 
     @Override
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     public void deleteTrainingById(Long trainingIdForDelete) {
         Training training = trainingRepository.findById(trainingIdForDelete)
                                               .orElseThrow(() -> new TrainingNotFoundException("Training not found!"));
+
+        Authentication auth = SecurityContextHolder.getContext()
+                                                   .getAuthentication();
+
+        boolean isUser = auth.getAuthorities()
+                             .stream()
+                             .anyMatch(a -> a.getAuthority()
+                                             .equals("ROLE_USER"));
+
+        if (isUser && !training.getUserId()
+                               .equals(getCurrentUserId())) {
+            throw new AccessDeniedException("You can only access your own trainings!");
+        }
+
         trainingRepository.deleteById(trainingIdForDelete);
     }
 
     @Override
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     public void addSetToTraining(Long trainingId, Integer reps) {
         Training training = getTrainingById(trainingId);
